backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
requests==2.32.3
beautifulsoup4==4.12.3
openai==1.37.1
pydantic==2.7.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
Pillow==10.3.0
qrcode==7.4.2
